backend/requirements.txt
fastapi
uvicorn
sqlalchemy
bcrypt
pyjwt
pyotp
qrcode
pydantic
